offshoot/captcha/requirements.txt
Flask==1.1.2
muggle-ocr==1.0.3
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0